BOURNE-1686: Avoid fetching all clients when deleting managed clients
When "remote state" is enabled for clients (the default), Keycloak will only delete clients that were listed in the config tool's state. This prevents it from deleting clients that it didn't create.

When deciding what to delete, we can just fetch the clients that were already in the state, rather than listing all the clients in the realm. This should be a much smaller list, and prevents issues where the server cuts off the response when the list is too large.

(cherry picked from commit 0376b11e48a958a70a71192aa7a13b6ee8b61e78)